feat(web): give the Channels tab a page-style layout
The standalone tab renders a page-level heading above the channel
accordion card, matching the sibling About Assistant tabs, instead of
reusing the Contacts detail's boxed "Channels" card. The accordion rows
and their confirm dialogs move into AssistantChannelsList, which both
mount points compose; AssistantChannelsDetail keeps the identity header
and card chrome for the Contacts pane. The ?setup= param read/consume
is shared via useSetupChannelParam.

Part of LUM-2696.
