+++ clients/web/src/domains/contacts/channels-page.tsx
@@ -1,9 +1,12 @@
 import { useQueryClient } from "@tanstack/react-query";
 import { useCallback, useState } from "react";
 
-import { AssistantChannelsDetail } from "@/domains/contacts/components/assistant-channels-detail";
+import { DetailCard } from "@/components/detail-card";
+import { AssistantChannelsList } from "@/domains/contacts/components/assistant-channels-list";
 import { GenerateInviteLinkDialog } from "@/domains/contacts/components/generate-invite-link-dialog";
+import { ShareConnectionLinkButton } from "@/domains/contacts/components/share-connection-link-button";
 import { useAssistantChannels } from "@/domains/contacts/hooks/use-assistant-channels";
+import { useSetupChannelParam } from "@/domains/contacts/hooks/use-setup-channel-param";
 import { contactsGetQueryKey } from "@/generated/daemon/@tanstack/react-query.gen";
 import { useAssistantFeatureFlagStore } from "@/stores/assistant-feature-flag-store";
 import { useAssistantIdentityStore } from "@/stores/assistant-identity-store";
@@ -16,9 +19,10 @@ export interface ChannelsPageProps {
 /**
  * Channels settings — the Slack/Telegram/Phone accordion where the guardian
  * manages how and where the assistant can be reached. Rendered as its own tab
- * in the About Assistant nav (`/assistant/channels`). The same surface also
- * renders inside the Contacts page's assistant detail view; both compose
- * `useAssistantChannels` + `AssistantChannelsDetail`.
+ * in the About Assistant nav (`/assistant/channels`): a page-level heading
+ * above the channel list, like the sibling tabs. The Contacts page's
+ * assistant detail renders the same list boxed as a card
+ * (`AssistantChannelsDetail`); both compose `useAssistantChannels`.
  */
 export function ChannelsPage({
   assistantId,
@@ -27,10 +31,11 @@ export function ChannelsPage({
   const a2aChannel = useAssistantFeatureFlagStore.use.a2aChannel();
   const identityName = useAssistantIdentityStore.use.name();
   const queryClient = useQueryClient();
+  const setupChannel = useSetupChannelParam();
 
   const [inviteDialogOpen, setInviteDialogOpen] = useState(false);
 
-  const assistantName = identityName ?? "your assistant";
+  const displayName = identityName?.trim() || "your assistant";
 
   const channelsController = useAssistantChannels({
     assistantId,
@@ -51,13 +56,22 @@ export function ChannelsPage({
   }, [queryClient, assistantId]);
 
   return (
-    <div className="mx-auto w-full max-w-3xl">
-      <AssistantChannelsDetail
-        assistantName={assistantName}
-        showIdentityCard={false}
-        onGenerateInviteLink={a2aChannel ? handleOpenInviteLink : undefined}
-        {...channelsController}
+    <div className="flex flex-col gap-6">
+      <DetailCard
+        showBorder={false}
+        title="Channels"
+        subtitle={`Manage where ${displayName} can be reached.`}
       />
+
+      <DetailCard>
+        <AssistantChannelsList
+          assistantName={displayName}
+          initialExpandedChannel={setupChannel}
+          {...channelsController}
+        />
+      </DetailCard>
+
+      {a2aChannel ? <ShareConnectionLinkButton onClick={handleOpenInviteLink} /> : null}
 
       <GenerateInviteLinkDialog
         open={inviteDialogOpen}
